Bring Cell in line with generic api.
Before we were simply updating the p variable for a cell directly.  This
is OK since the only thing that has access to the cell is the floor
class.  However, using update brings a unified api to the cell class and
removes the implementation details about how a cell gets updated from
teh floor class.  These classes are tightly coupled.  This helps break
apart some of that coupling